requirements.txt
opencv-python
onnxruntime
paho-mqtt